test: share canonical path identity helper
